split up sessionUtils, updated imports, del sessionUtils
