+++ server/routes/posts.js
@@ -2,7 +2,7 @@ const { PrismaClient } = require("../generated/prisma");
 const router = require("express").Router();
 const STATUS_CODES = require("../statusCodes");
 const { QUICKTIME, MOV, COMMENT, CREATE } = require("../utils/constants");
-const { recalculateInteractionAverages } = require("../utils/sessionUtils");
+const recalculateInteractionAverages = require("../utils/sessions/recalculateInteractionAverages").default;
 const scorePosts = require("../utils/postRecommendations/scorePosts").default;
 const { uploadFile, deleteFile } = require("../utils/googleCloudStorageUtils");
 const { getVideoDurationInSeconds } = require("get-video-duration");
